+++ CHANGELOG.md
@@ -12,6 +12,10 @@ All notable changes to this project will be documented in this file.
   When is `true` the blockchain requests use an archive database with full blockchain history
   but reduced field set.
 
+- `transactions_by_lt` query is added to `blockchain.account` to paginate account transactions by logical time
+
+- `thread` parameter in `blockchain.block_by_seq_no` and `blockchain.blocks` is deprecated. `shard` parameter should be used instead
+
 ## [0.60.1] - 2023-06-21
 
 ### Fixed